tests/wpt/tests/content-security-policy/reporting/report-clips-sample.https.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/content-security-policy/support/testharness-helper.js"></script>
  <meta http-equiv="Content-Security-Policy"
        content="require-trusted-types-for 'script'; trusted-types default">
</head>
<body>
  <script>
  const trimmedSampleLength = 40;
  const evalSample = "evil = '1234567890123456789012345678901234567890;'";
  const trimmedEvalSample = evalSample.substring(0, trimmedSampleLength);
  promise_test(t => {
    let evil = false;
    assert_throws_js(EvalError, _ => {
      eval(evalSample);
    });
    assert_false(evil);
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
      assert_equals(e.sample, `eval|${trimmedEvalSample}`);
    }));
  }, `Unsafe eval violation sample is clipped to ${trimmedSampleLength} characters.`);

  promise_test(t => {
    let evil = false;
    assert_throws_js(EvalError, _ => {
      eval?.(evalSample);
    });
    assert_false(evil);
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
      assert_equals(e.sample, `eval|${trimmedEvalSample}`);
    }));
  }, `Unsafe indirect eval violation sample is clipped to ${trimmedSampleLength} characters.`);

  const functionBody = "return '1234567890123456789012345678901234567890';";
  const sampleWithoutFunctionPrefix = `(a,b\n) {\n${functionBody}\n}`;

  promise_test(t => {
    assert_throws_js(EvalError, _ => {
      new Function("a", "b", functionBody);
    });
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
      assert_equals(e.sample, `Function|${sampleWithoutFunctionPrefix.substring(0, trimmedSampleLength)}`);
    }));
  }, "Function constructor - the other kind of eval - is clipped.");

  promise_test(t => {
    assert_throws_js(EvalError, _ => {
      const AsyncFunction = async function() {}.constructor
      new AsyncFunction("a", "b", functionBody);
    });
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
      assert_equals(e.sample, `Function|${sampleWithoutFunctionPrefix.substring(0, trimmedSampleLength)}`);
    }));
  }, "Async Function constructor is also clipped.");

  promise_test(t => {
    assert_throws_js(EvalError, _ => {
      const GeneratorFunction = function*() {}.constructor
      new GeneratorFunction("a", "b", functionBody);
    });
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
        assert_equals(e.sample, `Function|${sampleWithoutFunctionPrefix.substring(0, trimmedSampleLength)}`);
    }));
  }, "Generator Function constructor is also clipped.");

  promise_test(t => {
    assert_throws_js(EvalError, _ => {
      const AsyncGeneratorFunction = async function*() {}.constructor
      new AsyncGeneratorFunction("a", "b", functionBody);
    });
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
        assert_equals(e.sample, `Function|${sampleWithoutFunctionPrefix.substring(0, trimmedSampleLength)}`);
    }));
  }, "AsyncGenerator Function constructor is also clipped.");

  promise_test(t => {
    const a = document.createElement("a");
    assert_throws_js(TypeError, _ => {
      a.innerHTML = "1234567890123456789012345678901234567890xxxx";
    });
    assert_equals(a.innerHTML, "");
    return waitUntilCSPEventForTrustedTypes(t).then(t.step_func_done(e => {
      assert_equals(e.sample, "Element innerHTML|1234567890123456789012345678901234567890");
    }));
  }, "Trusted Types violation sample is clipped to 40 characters excluded the sink name.");
  </script>
</body>
</html>
