LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-clips-sample.https-expected.txt


FAIL Unsafe eval violation sample is clipped to 40 characters. assert_throws_js: function "_ => {
      eval("evil = '1234567890123456789012345678901234567890';");
    }" did not throw
FAIL Function constructor - the other kind of eval - is clipped. assert_throws_js: function "_ => {
      new Function("a", "b", "return '1234567890123456789012345678901234567890';");
    }" did not throw
FAIL Trusted Types violation sample is clipped to 40 characters excluded the sink name. assert_equals: expected "Element innerHTML|1234567890123456789012345678901234567890" but got "Element innerHTML|1234567890123456789012"
